+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## 21.11.0
+- ! Minor breaking change ! Multiple values now have a default limit adjustable at initialization:
+    - Maximum size of all string keys is now 128 characters by default.
+    - Maximum size of all values in key-value pairs is now 256 characters by default.
+    - Maximum amount of segmentation in one event is mow 30 key-value pairs by default.
+    - Maximum amount of breadcrumbs that can be recorded at once is now 100 by default.
+    - Maximum stack trace lines per thread is now 30 by default.
+    - Maximum stack trace line length is now 200 by default. 
 - Increased the default max event batch size to 100.
 
 ## 20.11